[#4] refactor: 생성 파트 컴포넌트, hook 분리 - 컴포넌트 관심사에 맞게 분리 - 생성 로직 custom hook으로 분리
